Clean up some legacy code and move mock Page defaults to test-use only
