+++ test/assets/frontend/behavior_list/behavior_list_spec.tsx
@@ -3,9 +3,9 @@ import * as TestUtils from 'react-addons-test-utils';
 import BehaviorList, {BehaviorListProps} from '../../../../app/assets/frontend/behavior_list/index';
 import BehaviorGroup from '../../../../app/assets/frontend/models/behavior_group';
 import BehaviorGroupCard from '../../../../app/assets/frontend/behavior_list/behavior_group_card';
-import Page from '../../../../app/assets/frontend/shared_ui/page';
 import {BehaviorVersionJson} from "../../../../app/assets/frontend/models/behavior_version";
 import {TriggerType} from "../../../../app/assets/frontend/models/trigger";
+import {getPageRequiredProps} from "../shared_ui/page_spec";
 
 const absoluteUrl = () => "https://nope/";
 
@@ -207,7 +207,7 @@ describe('BehaviorList', () => {
   function createBehaviorList(config: BehaviorListProps) {
     const footer = TestUtils.renderIntoDocument(<Footer/>) as Footer;
     return TestUtils.renderIntoDocument(
-      <BehaviorList {...config} {...Page.requiredPropDefaults()} onRenderFooter={footer.renderFooter} />
+      <BehaviorList {...config} {...getPageRequiredProps()} onRenderFooter={footer.renderFooter} />
     ) as BehaviorList;
   }
